Move fetch to composables and rename it to DataLoaderService
+ Misc improvements for useDataLoader (prev. useFetch) along the way :
  - Rename name to metric in useDataLoader
  - Replace watch by watchEffect in useDataLoader
  - We actually don't need to get oldValue, and we want loadData to be
  executed immediately. So using watchEffect is simpler.
  - Simplify useDataLoader by inlining the loadData function

Also we make sure DistributionGrid is also using DataLoaderService
