+++ powa/static/js/components/dynamic/config/AllCollectorsDetail.vue
@@ -75,7 +75,7 @@
 </template>
 
 <script setup>
-import { useFetch } from "@/utils/fetch.js";
+import { useDataLoader } from "@/composables/DataLoaderService.js";
 
 const props = defineProps({
   config: {
@@ -86,5 +86,5 @@ const props = defineProps({
   },
 });
 
-const { loading, data: collector } = useFetch(props.config.name);
+const { loading, data: collector } = useDataLoader(props.config.name);
 </script>
