+++ powa/static/js/components/dynamic/database/WizardThisDatabase.vue
@@ -22,7 +22,7 @@
 </template>
 
 <script setup>
-import { useFetch } from "@/utils/fetch.js";
+import { useDataLoader } from "@/composables/DataLoaderService.js";
 
 const props = defineProps({
   config: {
@@ -33,5 +33,5 @@ const props = defineProps({
   },
 });
 
-const { loading, data } = useFetch(props.config.name);
+const { loading, data } = useDataLoader(props.config.name);
 </script>
