+++ powa/static/js/components/dynamic/database/query/QueryExplains.vue
@@ -36,7 +36,7 @@
 </template>
 
 <script setup>
-import { useFetch } from "@/utils/fetch.js";
+import { useDataLoader } from "@/composables/DataLoaderService.js";
 import { formatSql } from "@/utils/sql.js";
 import _ from "lodash";
 
@@ -49,5 +49,5 @@ const props = defineProps({
   },
 });
 
-const { loading, data: plans } = useFetch(props.config.name);
+const { loading, data: plans } = useDataLoader(props.config.name);
 </script>
