+++ powa/static/js/components/dynamic/database/query/QueryIndexes.vue
@@ -97,7 +97,7 @@
 </template>
 
 <script setup>
-import { useFetch } from "@/utils/fetch.js";
+import { useDataLoader } from "@/composables/DataLoaderService.js";
 import { formatSql } from "@/utils/sql.js";
 
 const props = defineProps({
@@ -109,5 +109,5 @@ const props = defineProps({
   },
 });
 
-const { loading, data } = useFetch(props.config.name);
+const { loading, data } = useDataLoader(props.config.name);
 </script>
